soft-sme-backend/Aiven.ai/AI SQL bot/multi-agent-postgres-data-analytics-main/api-server/requirements.txt
Flask==3.0.0
google-genai
psycopg2-binary
python-dotenv
